Add blog.json support and setup routing
Adds a new blog.json format which belongs in the Dropbox folder. This contains all the blog files to load, together with metadata - eg. route, publishdate, state, etc. It no longer finds all files in the Dropbox folder. The route entered is also used to look up with blog post we're referring to.