Add support for 'featured' posts
We now have a ViewModelBase abstract class and a concrete type for each. The data that goes into the shared view (eg. all pages) goes in this base ViewModelBase class, and this is the @model definition in the shared view and shared partial views. Then the more specific views can @model the more relevant derived type which the controller passed.