Start work on tag support
This is work in progress, and starts adding a many-to-many relationship between the posts and the tags. Unfortunately EF Core doesn't support many to many properly yet, so we need to manually create an interim entity (hence the BlogPostTag class). Note that this tag work has been suspended in favor of the Dropbox HTTP implementation, so will come back to this once that's done.