Add 'Id' field to Blog.json
Previously I was using the 'Title' as the main key to check things like if a post already existed in the database when doing an incremental sync. This was a pain because it was harder to change the title for an existing post. So I'm adding an Id field to the blog.json file which I set manually. The ID then pushed to Sql Server just uses this instead of creating a new Guid for new entries.