If post is draft, then use its GUID ID as the draft route
Previously I manually set the route in blog.json to a temp path (I happened to just generate a new GUID for this, but it could have been anything). Now, it'll always use the ID as the route for draft posts. This makes more sense now that we're explicitly having the GUID ID mentioned in blog.json anyway (see recent change 4c4dc0110c).