Fix recently introduced issue breaking about and my talks pages
These pages are marked as draft so they don't appear in the listing pages. So when I added the functionality to automatically make draft posts use their GUID ID for the route instead - this broke these two pages. A better solution will be to add additional properties to indicate whether posts are listing items or not - but for now this is just a quick fix.