Comment out ServiceMonitor in web.yaml
I get the following error when Azure Pipelines does the kubectl apply:

"error: the namespace from the provided object "monitoring" does not match the namespace "blog". You must pass '--namespace=monitoring' to perform this operation."

I suspect they're using on old version on Kubectl - as a google finds a few people complaining about the namespace in the yaml not properly overridding. As I don't use the monitoring stuff anyway at the moment - I'm just going to comment it out. Because this was failing - it was then stopping the tasks.yaml from being applied.